extracted checking for a new version into UpdateKobalt; semantics did not change
