+++ src/main/kotlin/com/beust/kobalt/app/UpdateKobalt.kt
@@ -1,8 +1,10 @@
 package com.beust.kobalt.app
 
-import com.beust.kobalt.misc.GithubApi
-import com.beust.kobalt.misc.KobaltWrapperProperties
+import com.beust.kobalt.api.Kobalt
+import com.beust.kobalt.misc.*
 import com.beust.kobalt.wrapper.Main
+import java.io.File
+import java.util.concurrent.TimeoutException
 import javax.inject.Inject
 
 /**
@@ -20,5 +22,25 @@ public class UpdateKobalt @Inject constructor(val github: GithubApi, val wrapper
      */
     fun downloadKobalt() {
         Main.main(arrayOf("--download", "--no-launch"))
+    }
+
+    fun checkForNewVersion(latestVersionString: String) {
+        try {
+            val latestVersion = Versions.toLongVersion(latestVersionString)
+            val current = Versions.toLongVersion(Kobalt.version)
+            val distFile = File(KFiles.joinDir(KFiles.distributionsDir, latestVersionString))
+            if (latestVersion > current) {
+                if (distFile.exists()) {
+                    log(1, "**** Version $latestVersionString is installed")
+                } else {
+                    listOf("", "New Kobalt version available: $latestVersionString",
+                            "To update, run ./kobaltw --update", "").forEach {
+                        log(1, "**** $it")
+                    }
+                }
+            }
+        } catch(ex: TimeoutException) {
+            log(2, "Didn't get the new version in time, skipping it")
+        }
     }
 }
